docker/web/requirements.txt
django-storages[google]==1.7.2
google-auth==1.6.3
google-cloud-container==0.3.0
gunicorn==19.9.0
psycopg2==2.7.7
sentry-sdk==0.12.3
kubernetes==10.0.1
